feat(ee): move usage-alert workflow to ee/alerts
